fix: sync venv writes to wrong remote path
Previously, `jhack sync --venv` would incorrectly sync venv packages to the root of the venv rather than nested under the correct lib folder (eg: synced to ./venv/my-package instead of ./venv/lib/python3.12/site-packages/my-package).  This was caused by parsing the path of the changed file on `site-packages` which in turn removed some of the path information (lib/python3.12).

This commit fixes the issue by providing the low level functions with the local venv root path rather than a boolean `venv`.  This makes it easier to find the portable part of the package relative paths.